% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{dropdown_values}
\alias{dropdown_values}
\title{Scraping opentransportmap}
\usage{
dropdown_values(nuts = NULL, remDr = NULL)
}
\arguments{
\item{country}{Name of the country for which the map should be downloaded}

\item{browser}{Name of the browser used to download the data}
}
\description{
This function downloads every .zip file for a given country from the opentransportmap website.
Right now, the funciton should be use with Firefox only, and the files are downloaded in the default download directory of the local computer.
There are a few preliminary steps to carry out before using it, please refer to the document XXX.
}
